Two way system for forkXphilo action (#4)
* clean up the main func

The cleanup, should, not make any changes to
the structure of the program or the output

* add a delay to phi. threads

By adding a slow implementation of fibonacci
and calling it, the threads are spread out over
time, and thus don't deadlock

* change the order of fork and philo creation

The purpose is to make the forks ready before the
philofosers arrive. Then the threads dont't need to
be spaced out in time, using fibo, as before.

build on #2

* use fmt / lint on files

* change the fork to a active concurent thread

With this commit, the fork function, will
do something and not stand still.
It will also introduce a eating time system
to seperate the threads out in time.